server/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg==3.1.12
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
